requirements.txt
alpaca-py
toml
finnhub-python